refactor (database): Restructer fake db file to use {} instead of []
Our fake db file used to contain an array of room objects, each of which contained an array of furniture objects.
Now the file contains an object with a bunch of room names as keys, and each value is the room object. Each of
those objects contains a furniture key, with an object value that has furniture names for keys.