test: Write specs for deeplyHasKeyOrProp function
Many of our tests would benefit from the ability to search and determine if an object has a key or value somewhere inside. If the test specifies the exact structure of the object (e.g. expect(rooms.bathroom.furniture.toilet.cost).to.equal(100) ), the test will be brittle - it will fail if we change how we structure our data.

The deeplyHasKeyOrProp function will allow us to test if an object has a particular key or prop, and we don't need to know the structure of the object in advance.