feat: Add ability to mark projects private/public
We kind of had this feature before, but we didn't persist anything to the database. Now the checkbox updates the database and the redux state.

I also moved the button into the RoomsList, out of the navbar, where it was a bit out of place.

The checkbox reflects the current public/private status of the project.