feat(frontend): [WIP] Partially refactor furniture table
The end goal is to allow the user to view a table of furniture, and also click to edit/submit any row. I made some initial progress on this earlier, and it superficially appeared that I was further then than I am now, but this current refactor is meant to solve some problems I was running into in my previous attempt. The new version, when complete, should take care of setting the correct values in the form after a user cancels or saves, and should also allow the user to use the table sort functionality without exiting out of edit mode for all rows.