feat(frontend): Allow users to create furniture with no cost
Previously, a furniture with no cost broke our site in a few ways. Now, the budget graph displays correctly, and the rooms with zero cost are no longer deleted by the database. Yay!

Next I'll have to work on preventing our price placeholder ('literally priceless') from displaying anywhere on the site. I'll probably want to have the action that fetches the data from the db to strip those out before passing the data on to the reducer.